Fix UFuncTypeError in surrogate tests and add audit suite.
- Explicitly cast input data to float in `MannKS/_surrogate.py` to prevent type errors when applying censoring multipliers to integer arrays.
- Add `tests/audit_v060/` containing a comprehensive regression suite for v0.6.0 features:
    - `test_censored_surrogates.py`: Verifies correct propagation of censoring flags.
    - `test_surrogate_integrity.py`: Checks spectral and amplitude conservation.
    - `test_seasonal_alignment.py`: Ensures correct argument slicing in seasonal tests.
    - `test_power_analysis.py`: Validates power calculation logic.
    - `test_edge_cases.py`: Covers constant data and input validation.

+++ MannKS/_surrogate.py
@@ -330,7 +330,7 @@ def surrogate_test(
     notes = []
 
     # Apply imputation if censored data is present
-    x_eff = x_arr.copy()
+    x_eff = x_arr.copy().astype(float)
     if np.any(censored):
         # We need to distinguish between left and right censoring if possible
         # Typically 'cen_type' handles this: 'lt', 'gt', 'not'
